Split screen out of test scene
